pyproject.toml
[tool.poetry]
name = "whatsapp-pdf-assistant"
version = "0.1.0"
description = "WhatsApp PDF Assistant"
authors = ["Yarin Twito <yarin.twito@runi.ac.il>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<=3.13.3"
fastapi = ">=0.100.0,<=0.109.0"
uvicorn = ">=0.23.0,<=0.27.0"
httpx = "^0.25.0"
pytest = "^7.4.3"
python-multipart = "^0.0.6"
pypdf = "^3.17.0"
langchain = "^0.0.330"
openai = "^1.3.0"
Pillow = "^10.0.0"        # For image processing
PyMuPDF = "^1.23.0"       # For PDF to image conversion

[tool.poetry.group.dev.dependencies]
mypy = "^1.7.0"
flake8 = "^6.1.0"
black = "^23.11.0"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"